+++ src/ocaml_lsp.ml
@@ -44,6 +44,7 @@ module Experimental_capabilities = struct
     ; handleInferIntf : bool
     ; handleTypedHoles : bool
     ; handleTypeEnclosing : bool
+    ; handleConstruct : bool
     }
 
   let default =
@@ -51,6 +52,7 @@ module Experimental_capabilities = struct
     ; handleInferIntf = false
     ; handleTypedHoles = false
     ; handleTypeEnclosing = false
+    ; handleConstruct = false
     }
 
   (** Creates [t] given a JSON of form [{ 'handleSwitchImplIntf' : true, .... }] *)
@@ -65,10 +67,12 @@ module Experimental_capabilities = struct
       let handleInferIntf = has_capability "handleInferIntf" in
       let handleTypedHoles = has_capability "handleTypedHoles" in
       let handleTypeEnclosing = has_capability "handleTypeEnclosing" in
+      let handleConstruct = has_capability "handleConstruct" in
       { handleSwitchImplIntf
       ; handleInferIntf
       ; handleTypedHoles
       ; handleTypeEnclosing
+      ; handleConstruct
       }
     with Jsonoo.Decode_error err ->
       show_message
@@ -236,3 +240,5 @@ let can_handle_typed_holes t = t.experimental_capabilities.handleTypedHoles
 
 let can_handle_type_enclosing t =
   t.experimental_capabilities.handleTypeEnclosing
+
+let can_handle_construct t = t.experimental_capabilities.handleConstruct