+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Make it a warning if ocamlc is missing (#1642)
 - Add `1.18.0`, `1.19.0` and `1.20.0~5.3preview` to the list of known versions
   of ocamllsp (#1644)
+- Add `ocaml.construct` to construct an expression from a typedhole. (#1646)
 
 ## 1.20.1